+++ packages/orq-rc/RELEASES.md
@@ -8469,3 +8469,13 @@ Based on:
 - [typescript v4.8.0-rc.39] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.39] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.39 - packages/orq-rc
+
+## 2026-04-21 12:51:02
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.40] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.40] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.40 - packages/orq-rc